deleted argilla evaluation repository
